+++ src/components/SelectDropdown/index.js
@@ -27,6 +27,7 @@ class SelectDropdown extends React.Component {
      * Dropdown items that can be selected from the select dropdown.
      */
     items: PropTypes.arrayOf(PropTypes.shape({
+      activatorLabel: PropTypes.string,
       description: PropTypes.string,
       label: PropTypes.node.isRequired,
       value: PropTypes.oneOfType([
@@ -117,6 +118,8 @@ class SelectDropdown extends React.Component {
       {['oui-form-bad-news']: this.props.displayError}
     );
 
+    const activatorLabel = selectedItem.activatorLabel || selectedItem.label;
+
     return (
       <Dropdown
         { ...(zIndex ? { zIndex } : {}) }
@@ -130,7 +133,7 @@ class SelectDropdown extends React.Component {
               testSection={ this.props.testSection }
               width="full">
               <div className="flex flex-align--center" data-track-id={ this.props.trackId }>
-                <span style={{overflow: 'hidden'}} className="flex flex--1">{ selectedItem.label }</span>
+                <span style={{overflow: 'hidden'}} className="flex flex--1">{ activatorLabel }</span>
                 <span className="push--left oui-arrow-inline--down" />
               </div>
             </Button>