+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file is similar to the format suggested by [Keep a CHANGELOG](https://githu
 ## Unreleased
 - [Feature] Add displayError prop to SelectDropdown component
 - [Patch] Add mousedown handler to Button component
+- [Patch] Add support for activatorLabel property in SelectDropdown item objects
 
 ## 31.12.3 - 2018-08-24
 ### Changed